Fix corner cases + cleanup + add more comments
After some more heavy usage I noticed that there is always a space after
the last separator when show_tab_type=0, which needs to be taken into
account.

Additionally, sometimes it would add a buffer to the tabline
that filled the whole remaining space, meaning that the ellipsis would
overflow. This is now taken into account when calculating whether a
buffer fits.